Working on integer calculation for aht10

+++ include/aht10.h
@@ -26,8 +26,16 @@
 ////////////////////////////////////////////////////////////////////////////////
 /// Globals
 ////////////////////////////////////////////////////////////////////////////////
-extern float aht10_temperature;             // Last read temperature (deg C)
-extern float aht10_humidity;                // Last read humidity (%)
+
+// Last read temperature (deg C)
+// Last two digits are after decimal point
+extern unsigned int aht10_temperature;
+
+
+// Last read humidity (%)
+// Last two digits are after decimal point
+extern unsigned int aht10_humidity;
+
 extern unsigned int aht10_ec;               // Current error code for AHT10
 extern bbi2c_transaction aht10_trans;       // Transaction var for AHT10
